+++ packages/instant-apps-components/CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## v1.0.0-beta.207
+
+### instant-apps-sign-in
+
+- Add sign in component. Requires `portal` and `oauthappid` props.
+
+### instant-apps-landing-page
+
+- Add sign in capability to landing page.
+
 ## v1.0.0-beta.206
 
 ### instant-apps-social-share